pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pyqt6>=6.8.1",
    "requests>=2.32.3",
]
